move jb x url
