+++ src/components/layout/Footer/components/FullFooter.tsx
@@ -1,9 +1,9 @@
 import Logo from '@/components/Logo'
-import { JB_X_URL } from '@/components/icon/XLogo'
 import { twMerge } from 'tailwind-merge'
 import { DISCORD_INVITE_URL } from '../Footer'
 import { CopyrightBanner } from './CopyrightBanner'
 import { LinkColProps, LinkColumn } from './LinkColumn'
+import { JB_X_URL } from '@/lib/constants/urls'
 
 export type FullFooterProps = {
   className?: string
