+++ src/lib/constants/urls.ts
@@ -1,0 +1,1 @@
+export const JB_X_URL = 'https://twitter.com/Juicebox_Money'
